feat(cli): add SPEX Tester agent for comprehensive quality and security checks fix: update version to 1.0.23 fix(benches): replace black_box with std::hint::black_box in db_bench fix(deny): change yanked policy to warn and add advisory ignore for paste fix(cbor): simplify integer mapping in canonicalization test

+++ spex-core/src/cbor.rs
@@ -296,7 +296,7 @@ mod tests {
         fn canonicalization_is_idempotent_for_integer_arrays(values in proptest::collection::vec(any::<i64>(), 0..128)) {
             let value = Value::Array(
                 values.into_iter()
-                    .map(|item| Value::Integer((item as i64).into()))
+                    .map(|item| Value::Integer(item.into()))
                     .collect()
             );
             prop_assert!(stable_roundtrip(&value));
